move LoopIterator into own file
